Create Escaping Literal Quotes in Strings
When defining a string, start and end with single or double quotes [' ' or " "] but sometimes, you may need a quote inside of your string  (a quote inside quotes, in simpler terms). In JavaScript, you can escape a quote from considering it as an end-of-string quote by placing a backslash (\) [the button below the backspace on your keyboard] in front of the quote.