Create Use Bracket Notation to Find the First Character in a String
Javascript, a lot in computer science actually, is zero-based indexed. What this means is, traditionally you were taught to start counting from one, but in zero-based indexing, you start your count from zero. In Javascript, to get the first character in a string, you place a zero between a bracket, next to the name of the string or array like this [0].